move main.tsx into client folder
